test case for Pool.close()
This is a really lame test case--it simply verifies that the call does not throw error. I had a hard time to come up with good test cases, because the effect of calling close() is kind of hidden from the application (in fact, the Pool is still usable after it's closed!!). I also could not reproduce the program hanging problem in the mocha setting pre-fix as mocha seems to clean up things differently than just running a program by invoking node (?).