Add Math.floor to ensure that we overwrite indexes.
There is a bug here that Math.floor() was not being called.
The code was overwriting the previous 100 instances but instead
of overwriting them they were inserting floating points keys into
the array as keys turning the array into a dictionary.

When we add endpoints to lb_pool, we loop over the array and just
add the first 100 items, which are the original ones and not the
looping over the dictionary keys we inserted.